Ordset meta-table for listings; cpu stats probe
When creating entries, a record is also added to an ordset table
which is then used for find_entries() and select(). This gives good
prefix search performance and also ordered output.

A cpu stats probe was added, exometer_cpu, which samples cpu_sup
and caches the most recent result (in the probe state). Sampling
interval is controlled using the standard probe option
'sample_interval'.

A probe can be created with the type notation {exometer_probe,Mod},
where Mod is the exometer_probe behavior callback module.