Add 'ad_hoc' exometer_entry type
Normally, exometer_entry types must be made known to exometer
before entries of such types can be created. However, if a few odd
metrics need a different definition, it is awkward to first have to
define a template for that metric, and then create it.

The ad_hoc type requires that `module' and `type' are specified
in the Options list. The `type' option may be parameterized, e.g.
`{histogram, Arg}', which will be treated the same as if a known
type had been parameterized in that way (i.e. `{type_arg, Arg}'
is added to the Options list). See test/exometer_tests.erl for an
example of how to use.