exometer:setopts/2 calls optional callback M:preprocess_setopts/5
This is initially to allow dynamic modification of function entries.
In order to change a function entry, the 'ref' value must be updated.
For consistency, 'arg' should change correspondingly, but none of
these things supported in a good way. Also, the setopts/4 callback
did not really have a proper chance to vet the changes.

An optional callback, M:preprocess_setopts(Name,Opts,Type,Ref,OldOpts)
was introduced, which - if exported - is called and is expected to
return a new Options list. This is then processed as before, with the
exception that 'ref' can now also be changed.

The exometer_function:preprocess_setopts/5 callback ensures that
'ref' and 'arg' are consistent. However, a user modifying a
function entry should use change only the 'arg' option, as it is
part of the interface, whereas the representation of 'ref' should
be considered opaque.

Even so, it is now possible to change 'ref' via setopts/2.